// types.ts
export interface PosterFormData {
  title: string;
  authors: string;
  affiliations: string;
  mainIdea: string;
}

export interface PosterData {
  title: string;
  introduction: string;
  methods: string;
  results: string;
  discussion: string;
  conclusion: string;
  references: string;
  acknowledgements: string;
}